Version 3
Latest version. The token in the script.js file is redacted for obvious reasons. This version has all of the improvements requested on the review section of Version 2 completed, like fixing formatting errors, HTML/CSS convention and accessibility checks being satisfied, as well as improving the interactability, now including a call of action and an about page. More details are seen in the test log excel.